+++ CHANGELOG.md
@@ -14,6 +14,14 @@
   supply certificates with the
   `Options::add_root_certificate` method.
 * `ServerInfo` has been made non-public.
+* `Options::set_disconnect_callback` is renamed
+  `Options::disconnect_callback`.
+* `Options::set_reconnect_callback` is renamed
+  `Options::reconnect_callback`.
+* `Options::set_close_callback` is renamed
+  `Options::close_callback`.
+* `Options::set_reconnect_delay_callback` is renamed
+  `Options::reconnect_delay_callback`.
 
 # 0.6.0